# openapi/legal-rag-openapi.yaml
openapi: 3.1.0
info:
  title: Legal Agentic RAG Internal Platform API
  version: 0.1.0
  description: Spec-first API for corpus ingestion, QA runtime, evaluation, gold annotation,
    and synthetic dataset generation.
servers:
- url: https://api.example.internal
...
paths:
...
  /v1/runs/{runId}:
    get:
      tags:
      - Runs
      summary: Get one inference run
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                    format: uuid
                  dataset_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  question_count:
                    type: integer
                  created_at:
                    type: string
                    format: date-time
                required:
                - run_id
                - dataset_id
                - status
                - question_count
                - created_at
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Inference run UUID
...
  /v1/eval/runs:
    post:
      tags:
      - Evaluation
      summary: Launch evaluation of a run against a gold dataset
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalRun'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                run_id:
                  type: string
                  format: uuid
                gold_dataset_id:
                  type: string
                  format: uuid
                scoring_policy_version:
                  type: string
                judge_policy_version:
                  type: string
              required:
              - run_id
              - gold_dataset_id
              - scoring_policy_version
              - judge_policy_version
  /v1/eval/runs/{evalRunId}:
    get:
      tags:
      - Evaluation
      summary: Get evaluation run status and metrics
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalRun'
      parameters:
      - name: evalRunId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Evaluation run UUID
  /v1/eval/runs/{evalRunId}/report:
    get:
      tags:
      - Evaluation
      summary: Get full evaluation report with question-level breakdown
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  eval_run:
                    $ref: '#/components/schemas/EvalRun'
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        question_id:
                          type: string
                        answer_score:
                          type: number
                        grounding_score:
                          type: number
                        telemetry_factor:
                          type: number
                        ttft_factor:
                          type: number
                        overall_proxy:
                          type: number
                        error_tags:
                          type: array
                          items:
                            type: string
                      required:
                      - question_id
                      - answer_score
                      - grounding_score
                      - telemetry_factor
                      - ttft_factor
                      - overall_proxy
                      - error_tags
                required:
                - eval_run
                - items
      parameters:
      - name: evalRunId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Evaluation run UUID
  /v1/eval/compare:
    post:
      tags:
      - Evaluation
      summary: Compare two evaluation runs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  left_eval_run_id:
                    type: string
                    format: uuid
                  right_eval_run_id:
                    type: string
                    format: uuid
                  metric_deltas:
                    type: object
                  question_deltas:
                    type: array
                    items:
                      type: object
                required:
                - left_eval_run_id
                - right_eval_run_id
                - metric_deltas
                - question_deltas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                left_eval_run_id:
                  type: string
                  format: uuid
                right_eval_run_id:
                  type: string
                  format: uuid
              required:
              - left_eval_run_id
              - right_eval_run_id
  /v1/eval/calibrate-judge:
    post:
      tags:
      - Evaluation
      summary: Calibrate or benchmark free-text judge prompts/models
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                required:
                - job_id
                - status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                gold_dataset_id:
                  type: string
                  format: uuid
                sample_size:
                  type: integer
                  minimum: 10
                judge_policy_version:
                  type: string
              required:
              - gold_dataset_id
              - sample_size
              - judge_policy_version
...
components:
  schemas:
...
    EvalRun:
      title: EvalRun
      description: Evaluation run against a gold dataset.
      type: object
      additionalProperties: false
      properties:
        eval_run_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        gold_dataset_id:
          type: string
          format: uuid
        scoring_policy_version:
          type: string
        judge_policy_version:
          type: string
        status:
          type: string
          enum:
          - queued
          - running
          - completed
          - failed
        metrics:
          type: object
          additionalProperties: false
          properties:
            answer_score_mean:
              type: number
            grounding_score_mean:
              type: number
            telemetry_factor:
              type: number
            ttft_factor:
              type: number
            overall_score:
              type: number
            source_precision:
              type: number
            source_recall:
              type: number
            source_f_beta:
              type: number
            no_answer_precision:
              type: number
            no_answer_recall:
              type: number
            p50_ttft_ms:
              type: integer
            p95_ttft_ms:
              type: integer
          required:
          - answer_score_mean
          - grounding_score_mean
          - telemetry_factor
          - ttft_factor
          - overall_score
          - source_precision
          - source_recall
          - source_f_beta
      required:
      - eval_run_id
      - run_id
      - gold_dataset_id
      - scoring_policy_version
      - judge_policy_version
      - status
      - metrics
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []